+++ src/Licensing/Models/LicenseKeyDataModel.cs
@@ -7,7 +7,7 @@ namespace Licensing.Models {
     public class LicenseKeyDataModel {
 
         public string Email { get; set; }
-        public float EpochTime { get; set; }
+        public double EpochTime { get; set; }
         public int Expiry { get; set; }
         public SubscriptionTier Subscription { get; set; }
